requirements.txt
pandas==1.3.3
numpy==1.21.2
gspread==5.4.0
google-auth==2.19.0
tqdm==4.64.1
retrying==1.3.4
yfinance==0.2.20
pandas-datareader==0.10.0
